feat: Implement mobile controls and settings for Tetris game
- Added HoldPanel component to display held tetromino.
- Introduced MobileControls component for touch-based game controls.
- Created MobileSettings component for game settings including theme and music options.
- Developed NextPiecesPanel to show upcoming tetrominos.
- Implemented PieceRenderer for rendering tetrominos based on their shape and color.
- Updated game logic to handle piece holding, dropping, and rotation.
- Enhanced game state management with hooks for mobile responsiveness and theme handling.
- Defined game constants and types for better structure and maintainability.
