+++ src/app/components/game/index.ts
@@ -1,0 +1,10 @@
+// Game components
+export { GameBoard } from "./GameBoard";
+export { GameStats } from "./GameStats";
+export { HoldPanel } from "./HoldPanel";
+export { NextPiecesPanel } from "./NextPiecesPanel";
+export { MobileControls } from "./MobileControls";
+export { MobileSettings } from "./MobileSettings";
+export { GameControls } from "./GameControls";
+export { GameAudio } from "./GameAudio";
+export { PieceRenderer } from "./PieceRenderer";
