+++ src/hooks/index.ts
@@ -1,0 +1,5 @@
+// Game hooks
+export { useTheme } from "./useTheme";
+export { useGameLogic } from "./useGameLogic";
+export { useGameControls } from "./useGameControls";
+export { useMobile } from "./useMobile";
